feat: add compare-report CLI subcommand
