+++ src/tool_eval_bench/compare_reports/__init__.py
@@ -1,0 +1,1 @@
+"""HTML comparison report generators for existing Markdown benchmark reports."""
